refactor: LLM판결 V2 CQRS 적용
