+++ src/main/java/store/lastdance/service/aijudgment/AiJudgmentV2Service.java
@@ -3,12 +3,10 @@ package store.lastdance.service.aijudgment;
 import store.lastdance.dto.aijudgment.CreateAiJudgmentRequestDTO;
 import store.lastdance.dto.aijudgment.AiJudgmentResponseDTO;
 
-import java.util.List;
 import java.util.UUID;
 
 public interface AiJudgmentV2Service {
     AiJudgmentResponseDTO judgeConflict(CreateAiJudgmentRequestDTO request, UUID userId);
     String toggleFeedback(UUID judgmentId, UUID userId, String type);
-    List<AiJudgmentResponseDTO> getAiJudgmentHistory(UUID userId);
     void deleteAiJudgment(UUID judgmentId, UUID userId);
 }
